backend/pyproject.toml
[project]
name = "decision-making-deep-research-assistant"
version = "0.1.0"
description = "Deep research assistant with LangGraph agent orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "yfinance>=0.2.40",
    "pinecone>=5.0.0",
    "sqlmodel>=0.0.21",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
    "openai>=1.50.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "greenlet>=3.3.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
